backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymilvus==2.4.7
marshmallow<4
openai==1.51.0
httpx==0.27.2
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
pydantic==2.9.2
tiktoken==0.8.0
